refactor: InvitationCode를 InvitationCodeGenerator로 대체
